구조 완전 변경
