+++ src/main/java/com/app/Main.java
@@ -1,4 +1,4 @@
-package src.main;
+package com.app;
 
 public class Main {
     public static void main(String[] args) {
