+++ src/main/java/com/app/MainController.java
@@ -1,13 +1,16 @@
-package src.main;
+package com.app;
 
-import src.screen.*;
-import src.screen.interfaces.Screen;
+
+import org.json.JSONObject;
+import com.screen.*;
+import com.screen.interfaces.Screen;
 
 import javax.swing.*;
 import java.awt.*;
+import java.io.BufferedReader;
+import java.io.FileReader;
 import java.util.HashMap;
 import java.util.Map;
-
 
 public class MainController extends JFrame {
     private CardLayout cardLayout;
@@ -35,7 +38,7 @@ public class MainController extends JFrame {
     }
 
     private void addScreen(String name, Screen screen) {
-        screens.put(name, screen); // Map에 화면 추가
+        screens.put(name, screen); // Ma
         mainPanel.add((Component) screen, name); // CardLayout에 화면 추가
     }
 
@@ -51,10 +54,31 @@ public class MainController extends JFrame {
         cardLayout.show(mainPanel, screenName);
     }
 
+    public JSONObject loadJson(String path) {
+
+
+
+        // 지정된 경로에서 JSON 파일을 로드
+        try (BufferedReader reader = new BufferedReader(new FileReader(path))) {
+            StringBuilder jsonText = new StringBuilder();
+            String line;
+            while ((line = reader.readLine()) != null) {
+                jsonText.append(line);
+            }
+
+            // JSON 텍스트를 JSONObject로 변환
+            return new JSONObject(jsonText.toString());
+
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+
     public void loadSettingData() {
     }
 
     public void saveData() {
-
     }
 }
