+++ src/main/java/com/item/interfaces/Item.java
@@ -1,4 +1,4 @@
-package src.item.interfaces;
+package com.item.interfaces;
 
 public interface Item {
