+++ src/main/java/com/screen/ResetScreen.java
@@ -1,22 +1,21 @@
-package src.screen;
+package com.screen;
 
-import src.screen.interfaces.Screen;
-import src.main.MainController;
+
+
+import com.app.MainController;
 
 import javax.swing.*;
 
-public class LoadScreen extends JPanel implements Screen {
+import com.screen.interfaces.Screen;
+
+public class ResetScreen extends JPanel implements Screen {
+
     private MainController mainController;
 
-    public LoadScreen(MainController mainController) {
+    public ResetScreen(MainController mainController) {
         this.mainController = mainController;
         initialize();
     }
-
-    private void loadData() {
-
-    }
-
     @Override
     public void initialize() {
 
@@ -26,7 +25,6 @@ public class LoadScreen extends JPanel implements Screen {
     public void showScreen() {
         setVisible(true);
     }
-
     @Override
     public void hideScreen() {
         setVisible(false);
